requirements.txt
Flask>=3.0.3
gunicorn=22.0.0
python-dotenv>=1.0.1
requests>=2.32.3
peewee>=3.17.6
psycopg2-binary>=2.9.9
playhouse
openai>=1.40.8
gevent>=24.2.1

